+++ multiqc/modules/isoseq3_cluster/__init__.py
@@ -1,0 +1,2 @@
+from __future__ import absolute_import
+from .isoseq3_cluster import MultiqcModule